requirements.txt
# Core dependencies
pyrogram>=2.0.106,<3.0.0
# On Windows, you might need to install tgcrypto from a pre-compiled wheel
# See: https://www.lfd.uci.edu/~gohlke/pythonlibs/#tgcrypto
tgcrypto>=1.2.5,<2.0.0; sys_platform != "win32"
python-dotenv>=1.0.0,<2.0.0

# Video processing
ffmpeg-python>=0.2.0,<0.3.0

# M3U8 handling
m3u8>=3.5.0,<4.0.0

# Networking
requests>=2.31.0,<3.0.0
aiohttp>=3.8.5,<4.0.0
aiofiles>=23.1.0,<24.0.0
jinja2>=3.1.2,<4.0.0
aiohttp-basicauth>=1.0.0,<2.0.0
markdown>=3.4.3,<4.0.0

# System monitoring
psutil>=5.9.5,<6.0.0

# Platform-specific dependencies
# Windows: tgcrypto should be installed manually from a pre-compiled wheel
# macOS: No additional dependencies
# Linux: No additional dependencies
